Week8_AI/requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-google-genai
chromadb==0.6.3
google-generativeai
streamlit==1.45.0
langsmith==0.3.42
python-dotenv==1.1.0
tiktoken==0.9.0